Fix HIP dispatch to pass 32-bit loop bound matching codegen
The HIP codegen emits 'int __n__' (32-bit) but the dispatch was passing
c_longlong (64-bit), which would cause incorrect argument passing on
real AMD hardware.

+++ src/pgc/runtime/hip_backend.py
@@ -161,7 +161,7 @@ class CompiledHIPKernel:
 
     def __call__(self, device_ptrs: list, loop_end: int):
         """Dispatch the HIP kernel."""
-        n_val = ctypes.c_longlong(loop_end)
+        n_val = ctypes.c_int(loop_end)
 
         arg_values = []
         for dptr in device_ptrs:
